rename plugin-odoo-snippets to plugin-odoo
