+++ packages/plugin-odoo/src/LinkButton.ts
