+++ packages/plugin-odoo/src/OdooBindings.ts
