+++ packages/plugin-odoo/src/OdooImageHtmlDomRenderer.ts
