+++ packages/plugin-odoo/src/OdooStructureNode.ts
