+++ packages/plugin-odoo/src/OdooStructureXmlDomParser.ts
